Extract Component-based gate functions into hwsim.logic module
Move all functions that instantiate Component (and_gate, or_gate,
any_gate, xor_gate, nxor_gate, isz_gate, mux_gate, not12, and12)
from hwsim.component into a new hwsim.logic module.

Update hwsim.arithmetic and tests to import from the new location.
